Rearchitecting to allow granular feature set
