--- inc/dashboard/api/namespace.php
@@ -1,151 +1,0 @@
-<?php
-/**
- * Altis Accelerate analytics API.
- *
- * @package altis/accelerate
- */
-
-namespace Altis\Analytics\Dashboard\API;
-
-use Altis\Analytics\Dashboard;
-use Altis\Analytics\Dashboard\Filter;
-use WP_REST_Request;
-use WP_REST_Response;
-use WP_REST_Server;
-
-/**
- * Bootstrap analytics API.
- *
- * @return void
- */
-function bootstrap() {
-	add_action( 'rest_api_init', __NAMESPACE__ . '\\register_endpoints' );
-}
-
-/**
- * Register APi routes.
- *
- * @return void
- */
-function register_endpoints() {
-	$date_args = [
-		'start' => [
-			'type' => 'string',
-			'format' => 'date-time',
-			'required' => true,
-		],
-		'end' => [
-			'type' => 'string',
-			'format' => 'date-time',
-			'required' => true,
-		],
-		'interval' => [
-			'type' => 'string',
-			'enum' => [
-				'day',
-				'hour',
-			],
-			'default' => 'day',
-		],
-	];
-
-	register_rest_route( Dashboard\API_NAMESPACE, 'stats', [
-		'method' => WP_REST_Server::READABLE,
-		'callback' => __NAMESPACE__ . '\\get_stats',
-		'permission_callback' => __NAMESPACE__ . '\\check_analytics_permission',
-		'args' => array_merge( [
-			'path' => [
-				'type' => 'string',
-			],
-		], $date_args ),
-	] );
-	register_rest_route( Dashboard\API_NAMESPACE, 'top', [
-		'method' => WP_REST_Server::READABLE,
-		'callback' => __NAMESPACE__ . '\\get_top',
-		'permission_callback' => __NAMESPACE__ . '\\check_analytics_permission',
-		'args' => array_merge( [
-			'type' => [
-				'type' => 'string',
-			],
-			'author' => [
-				'type' => 'number',
-			],
-			'search' => [
-				'type' => 'string',
-			],
-			'page' => [
-				'type' => 'number',
-				'default' => 1,
-			],
-		], $date_args ),
-	] );
-}
-
-/**
- * Check user permissions for viewing analytics data.
- *
- * @return bool
- */
-function check_analytics_permission() : bool {
-	return current_user_can( 'edit_posts' );
-}
-
-/**
- * Handle stats endpoint.
- *
- * @param WP_REST_Request $request The request object.
- * @return WP_Error|array
- */
-function get_stats( WP_REST_Request $request ) {
-	$start = strtotime( $request['start'] );
-	$end = strtotime( $request['end'] );
-
-	$filter = new Filter();
-	if ( $request['filter'] && is_array( $request['filter'] ) ) {
-		foreach ( $request['filter'] as $key => $value ) {
-			if ( property_exists( $filter, $key ) ) {
-				$filter->$key = $value;
-			}
-		}
-	}
-
-	return Dashboard\get_graph_data( $start, $end, $request['interval'], $filter );
-}
-
-/**
- * Handle top content endpoint.
- *
- * @param WP_REST_Request $request The request object.
- * @return WP_Error|WP_REST_Response
- */
-function get_top( WP_REST_Request $request ) {
-	$start = strtotime( $request['start'] );
-	$end = strtotime( $request['end'] );
-
-	$filter = new Filter();
-	if ( $request['search'] ) {
-		$filter->search = $request['search'];
-	}
-	if ( $request['type'] ) {
-		$filter->type = $request['type'];
-	}
-	if ( $request['user'] ) {
-		$filter->user = (int) $request['user'];
-	}
-	if ( $request['page'] ) {
-		$filter->page = (int) $request['page'];
-	}
-
-	$data = Dashboard\get_top_data( $start, $end, $filter );
-
-	if ( is_wp_error( $data ) ) {
-		return $data;
-	}
-
-	$response = new WP_REST_Response( $data['posts'], 200, [
-		'X-WP-Total' => $data['total'],
-		'X-WP-TotalPages' => $data['max_pages'],
-	] );
-
-	return $response;
-}
